requirements.txt
groq
python-dotenv
rich
Pillow
pygame
pygame_gui
